Add new option to auto save all buffers on navigation
Summary: Added a new option which, when set, causes all open and changed
buffers to be written when navigating away from a vim session. If auto
save is turned on and this option is turned off, only the active buffer
gets saved.

Test Plan:
1) Open up two tmux panes, one running vim and the other not
2) In the vim pane open several files and modify each of them without
explicitly writing any of them
3) navigate away from the tmux split containing vim
4) navigate back to the tmux split containing vim
5) write all files with ":wa"
6) see that nothing got written suggesting that every buffer was written
in step 3

1) enable new option
2) open two tmux splits
3) in one of the splits open vim without a file name. (just `vim` not
`vim <some-file-path>`)
4) enter text into the vim instance
5) navigate away from tmux pnae containing unnamed vim
6) see that no error is shown to user